Move song routes under /songs
